renaming new files
