+++ src/containers/control/view/Connect.js
