+++ src/routes.js
@@ -28,8 +28,8 @@ import WelcomeUserView from './containers/welcome-user'
 import AddObservationPage from './containers/observation/add'
 import EditObservationPage from './containers/observation/edit'
 import ViewObservationPage from './containers/observation/view'
-import ControlViewContainer from './containers/control/view/Component'
-import ControlAddContainer from './containers/control/add/Component'
+import ControlViewContainer from './containers/control/view/Connect'
+import ControlAddContainer from './containers/control/add/Connect'
 import App from './containers/app'
 import ObservationControlGridShow from './containers/observationcontrol/grid'
